perf: all core optimizations — Level 1-2 + kernel fusions
Summary of all optimizations implemented:

Level 1: SSM State BF16化
- GMEM BF16 storage for DeltaNet SSM states, kernel-internal FP32 compute
- B=1 no regression, B=128 throughput +42.6%

Level 2: Projection GEMV Merge
- FullAttn QKV merge: [12288+1024+1024, 5120] → single GEMV (16 layers × 2 = 32 launches saved)
- LinearAttn QKVZAB super-merge: [10240+6144+48+48, 5120] → single GEMV (48 layers × 3 = 144 launches saved)
- Init-time weight concatenation, original weights released (net zero memory)

Kernel Fusions:
- Fused QK_norm + RoPE: deinterleave + Q_norm + K_norm + partial RoPE → 1 kernel (32 launches saved)
- Fused RMSNorm + GEMV: input RMSNorm computed in GEMV SMEM, eliminates norm_out GMEM I/O (64 launches saved)
- GPU Sampling with Gumbel-Max fast path (from FlashInfer)

Other changes:
- WY chunkwise prefill for DeltaNet (Phase 16, 1.71x speedup for T≥4)
- Paged attention Split-K decode
- Streaming attention for 256K+ contexts
- SSD KV/SSM state offload
- Continuous batching + MTP speculative decoding
- L2 persistence hints
- Benchmark improvements (CSV output, batch mode)
- Serve: Ollama/OpenAI compatible HTTP API improvements
- TUI improvements

Performance (B=1 decode):
- Baseline: 227.57ms median, 4.39 tok/s, 225.2 GB/s
- Current: ~228.8ms ITL, ~217.7ms forward, ~224 GB/s
- ~240 launches/step saved total

Failed experiments (reverted):
- GDN SMEM caching: occupancy drop negated benefit
- Dual GEMV + SwiGLU fusion: block count reduction killed scheduling density (+4.6% regression)

+++ src/engine/cache_engine.cpp
@@ -117,7 +117,7 @@ void CacheEngine::store_prefix(
     const ops::KVCacheManager& kv_manager,
     const int* d_block_table,
     int num_blocks,
-    float** ssm_states,
+    __nv_bfloat16** ssm_states,
     __nv_bfloat16** conv_states,
     __nv_bfloat16* workspace,
     cudaStream_t stream)
@@ -216,7 +216,7 @@ int CacheEngine::retrieve_prefix(
     const int* tokens, int num_tokens,
     ops::KVCacheManager& kv_manager,
     std::vector<int>& out_block_table,
-    float** ssm_states,
+    __nv_bfloat16** ssm_states,
     __nv_bfloat16** conv_states,
     __nv_bfloat16* workspace,
     cudaStream_t stream)
@@ -312,7 +312,7 @@ int CacheEngine::retrieve_prefix(
 // SSM/Conv 状态拷贝
 // ---------------------------------------------------------------------------
 void CacheEngine::copy_state_to_host(CacheEntry& entry,
-                                     float** ssm_states,
+                                     __nv_bfloat16** ssm_states,
                                      __nv_bfloat16** conv_states,
                                      cudaStream_t stream) {
     int num_lin = params_.num_linear_attn_layers;
@@ -334,7 +334,7 @@ void CacheEngine::copy_state_to_host(CacheEntry& entry,
 }
 
 void CacheEngine::copy_state_from_host(const CacheEntry& entry,
-                                       float** ssm_states,
+                                       __nv_bfloat16** ssm_states,
                                        __nv_bfloat16** conv_states,
                                        cudaStream_t stream) {
     int num_lin = params_.num_linear_attn_layers;
