+++ src/engine/cache_engine.h
@@ -70,7 +70,7 @@ public:
         const ops::KVCacheManager& kv_manager,
         const int* d_block_table,
         int num_blocks,
-        float** ssm_states,
+        __nv_bfloat16** ssm_states,
         __nv_bfloat16** conv_states,
         __nv_bfloat16* workspace,
         cudaStream_t stream);
@@ -81,7 +81,7 @@ public:
         const int* tokens, int num_tokens,
         ops::KVCacheManager& kv_manager,
         std::vector<int>& out_block_table,
-        float** ssm_states,
+        __nv_bfloat16** ssm_states,
         __nv_bfloat16** conv_states,
         __nv_bfloat16* workspace,
         cudaStream_t stream);
@@ -103,10 +103,10 @@ public:
 
 private:
     void copy_state_to_host(CacheEntry& entry,
-                            float** ssm_states, __nv_bfloat16** conv_states,
+                            __nv_bfloat16** ssm_states, __nv_bfloat16** conv_states,
                             cudaStream_t stream);
     void copy_state_from_host(const CacheEntry& entry,
-                              float** ssm_states, __nv_bfloat16** conv_states,
+                              __nv_bfloat16** ssm_states, __nv_bfloat16** conv_states,
                               cudaStream_t stream);
 
     // 更新容量指标快照
